refactor(notifications): add a notificationContainer and a notification-directory
