+++ src/components/chatBox.js
@@ -10,7 +10,7 @@ import ScrollableInkTabBar from 'rc-tabs/lib/ScrollableInkTabBar'
 import ChatDialog from './chatDialog'
 import FriendsList from './friendsList'
 import GroupsList from './groupsList'
-import Notifications from './notifications'
+import Notifications from '../containers/notificationsContainer'
 
 import 'rc-tabs/assets/index.css'
 
@@ -62,11 +62,7 @@ export default function ChatBox (props) {
 
     {props.shouldDisplayNotifications
       ? <TabPane tab='Notifications' key='notifications'>
-        <Notifications
-          notifications={props.notifications}
-          onClick={console.warn.bind(console, 'TODO: Implement callbacks')}
-          onCancel={props.closeNotification}
-        />
+        <Notifications />
       </TabPane>
       : null
     }
